[v 1.0.0 ] [ 🍀 - Authentication 네이밍 및 매개변수 리팩토링 ]
