--- iTime/ProxyPackage/Sources/ProxyPackage/Protocols/ASAuthorizationContextProviding.swift
@@ -1,10 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by 이상헌 on 2023/11/02.
-//
-
-import AuthenticationServices
-
-public protocol ASAuthorizationContextProviding: ASAuthorizationControllerPresentationContextProviding {}
